Fix: set secureProxy to true, unset secure
Cookies need to have "sameSite=none" and "Secure" to be used
cross-site with the new Chrome update. Previously, setting
"secure: true" to cookie-session resulted in no cookies being set.

Heroku terminates SSL before it reaches this node server, and thus all
requests appear to be http requests to the node-server. From looking at
pillarjs/cookies, setting "secure: true" when the request is not secure
throws an error in the cookie creation process, and is likely why no
cookies were returned whenever "secure: true" was set.

On the other hand, setting secureProxy to true explicitly tells
the cookies pkg that the requests are proxied and are to be trusted.
Thus, no error is thrown and cookies are set as Secure.